requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
opencv-python==4.9.0.80
python-multipart==0.0.6
boto3==1.34.14
Pillow==10.2.0